[Loading] Dramatically reduced model, scene loading and engine startup time by generating mips on the GPU using a single dispatch [Memory] Improved GPU performance and reduced memory usage by allowing textures lower than 32 bit and/or 3 channels to be loaded [Pass_AMD_FidelityFX_SinglePassDowsnampler] Fixed a bug where when dowsampling, mips beyond 6, would remain empty due to a structured buffer issue (which is used as an atomic counter). [Stability] Fixed a crash that could occur when loading a world [Stability] Fixed a crash that would occur when cloasing the editor and using D3D11 [TAA] Fixed some issues and improved quality [ThirdParty] Updated premake to latest

+++ Runtime/Rendering/Renderer_Primitives.cpp
@@ -40,7 +40,7 @@ namespace Spartan
         uint32_t end = static_cast<uint32_t>(m_lines_depth_disabled_duration.size());
         for (uint32_t i = 0; i < end; i++)
         {
-            m_lines_depth_disabled_duration[i] -= delta_time;
+            m_lines_depth_disabled_duration[i] -= static_cast<float>(delta_time);
         
             if (m_lines_depth_disabled_duration[i] <= 0.0f)
             {
@@ -54,7 +54,7 @@ namespace Spartan
         end = static_cast<uint32_t>(m_lines_depth_enabled_duration.size());
         for (uint32_t i = 0; i < end; i++)
         {
-            m_lines_depth_enabled_duration[i] -= delta_time;
+            m_lines_depth_enabled_duration[i] -= static_cast<float>(delta_time);
         
             if (m_lines_depth_enabled_duration[i] <= 0.0f)
             {
